src/dockerProject/kafkaCollector/requirements.txt
kafka-python>=2.0.0